requirements - Copy.txt
# DERM-EQUITY: Equitable Skin Cancer Detection
# Python 3.9+ required

# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
pytorch-lightning==2.1.0
timm==0.9.12

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# Medical Imaging
pydicom==2.4.3
nibabel==5.1.0
SimpleITK==2.3.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Experiment Tracking
wandb==0.16.0
tensorboard==2.15.1

# Configuration
omegaconf==2.3.0
hydra-core==1.3.2
pyyaml==6.0.1

# Utilities
tqdm==4.66.1
rich==13.7.0
python-dotenv==1.0.0

# Demo Application
gradio==4.8.0
streamlit==1.28.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Fairness & Uncertainty
uncertainty-toolbox==0.1.1

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
